Add functioning group forms and added show page styling

+++ frontend/components/groups/show/header/group_show_header_nav.jsx
@@ -4,7 +4,7 @@ import NavTabs from './group_header_tabs';
 
 class GroupShowHeaderNav extends React.Component {
     memberButton() {
-        const group = this.props.group
+        const group = this.props.group;
 
         let button = <button onClick={() => this.props.createMember(group.id)}>
             Join this group
@@ -13,7 +13,8 @@ class GroupShowHeaderNav extends React.Component {
         let members = this.props.members;
 
         for (let id in members) {
-            if (members[id].user_id === this.props.currentUser.id) {
+            if (this.props.currentUser && 
+                members[id].user_id === this.props.currentUser.id) {
                 button = <button onClick={() => this.props.deleteMember(id)}>
                     Leave this group
                 </button>
@@ -23,23 +24,54 @@ class GroupShowHeaderNav extends React.Component {
         return button;
     }
 
+    displayMemberButton() {
+        const group = this.props.group;
+
+        if (this.props.currentUser && 
+            this.props.group.creator_id === this.props.currentUser.id) {
+            return <div className="not-displayed"></div>;
+        } else {
+            return this.memberButton()
+        }
+    }
+
     editGroupButton() {
         let group = this.props.group;
 
         let button = <div className="pink-btn-container" id="pink-btn-show-pg">
             <div className="pink-btn">
-                <Link to={`/groups/${group.id}/edit`}>Edit Group</Link>
+                <div id="show-edit">
+                    <Link to={`/groups/${group.id}/edit`}>Edit Group</Link>
+                </div>
             </div>
         </div>
 
-        if (group.creator_id !== this.props.currentUser.id) {
-            console.log(button)
-            console.log(group.creator_id);
-            console.log(this.props.currentUser.id)
-            button = "";
+        if (this.props.currentUser && 
+            group.creator_id !== this.props.currentUser.id) {
+            button = <div className="not-displayed"></div>;
         }
 
         return button
+    }
+
+    deleteGroupButton() {
+        let group = this.props.group;
+
+        // console.log(this.props)
+        let button = <div className="pink-btn-container" id="pink-btn-show-pg">
+            <div className="pink-btn">
+                <button onClick={() => this.props.deleteGroup(group.id)}>
+                    Delete Group
+                </button>
+            </div>
+        </div>
+
+        if (this.props.currentUser &&
+            group.creator_id !== this.props.currentUser.id) {
+            button = <div className="not-displayed"></div>
+        }
+
+        return button;
     }
 
     render() {
@@ -48,23 +80,26 @@ class GroupShowHeaderNav extends React.Component {
         return (
             <div className="show-navbar-container">
                 <div className="show-navbar">
-                    <Link to={`/groups/${group.id}`} id="nav-button">About</Link>
+                    {/* <Link to={`/groups/${group.id}`} id="nav-button">About</Link>
                     <Link to={`/groups/${group.id}`} id="nav-button">Events</Link>
                     <Link to={`/groups/${group.id}`} id="nav-button">Members</Link>
-                    <Link to={`/groups/${group.id}`} id="nav-button">Photos</Link>
-                    {/* <div className="tabs-container">
+                    <Link to={`/groups/${group.id}`} id="nav-button">Photos</Link> */}
+                    <div className="tabs-container">
                         <NavTabs 
                             group={group}
                             members={this.props.members}
                         />
-                    </div> */}
+                    </div>
 
-                    <div className="pink-btn-container" id="pink-btn-show-pg">
+                </div>
+                <div className="show-nav-btns">
+                    <div className="pink-btn-container">
                         <div className="pink-btn">
-                            {this.memberButton()}
+                            {this.displayMemberButton()}
                         </div>
                     </div>
                     {this.editGroupButton()}
+                    {this.deleteGroupButton()}
                 </div>
             </div>
         )
